Deleted unecessary files and renamed test
